Move concepts to their own base directive folder
