+++ src/scripts/app/cms/concepts/directives/conceptForm.html
@@ -1,5 +1,6 @@
 <dynamic-form template="conceptTemplate"
   ng-model="concept">
 </dynamic-form>
+<standards-selector model="concept"></standards-selector>
 <concepts-selector model="concept" levels="[1,2]"></concepts-selector>
 <button ng-click="processConceptForm()">Update Concept</button>
