+++ src/scripts/app/cms/grammarActivities/directives/index.js
@@ -8,24 +8,9 @@ angular.module('quill-grammar.cms.grammarActivities.directives', [
   require('../../../../services/lms/concept.js').name,
   require('../../../../services/lms/standard.js').name,
   require('../../../../services/lms/standard_level.js').name,
-  require('../../base_directives/index.js').name,
+  require('../../base_directives/concepts/index.js').name,
 ])
 .controller('GrammarActivityFormCtrl', require('./grammarActivityForm.controller.js'))
 .directive('grammarActivityForm', require('./grammarActivityForm.directive.js'))
-.filter('filterConcepts', function (_) {
-  return function (concepts, parentId) {
-    return _.where(concepts, {parent_id: parentId});
-  };
-})
-.filter('filterFBConcepts', function (_) {
-  return function (concepts, parentId) {
-    return _.reject(concepts, function (c) {
-      if (!c.concept_level_0) {
-        return false;
-      } else {
-        return c.concept_level_0.parent_id !== parentId;
-      }
-    });
-  };
-})
+
 ;
